shuffle classes around
