+++ openapi-codegen/core/src/main/scala/sttp/tapir/codegen/endpoints/StreamingImplementation.scala
@@ -1,0 +1,7 @@
+package sttp.tapir.codegen.endpoints
+
+sealed trait StreamingImplementation
+object Akka extends StreamingImplementation
+case class FS2(effectType: String = "cats.effect.IO") extends StreamingImplementation
+object Pekko extends StreamingImplementation
+object Zio extends StreamingImplementation
