+++ openapi-codegen/core/src/test/scala/sttp/tapir/codegen/ClassDefinitionGeneratorSpec.scala
@@ -1,12 +1,15 @@
 package sttp.tapir.codegen
 
 import sttp.tapir.codegen.dedup.PackageReuseContext
+import sttp.tapir.codegen.endpoints.{EndpointGenerator, FS2}
 import sttp.tapir.codegen.json.JsonSerdeLib
 import sttp.tapir.codegen.openapi.models.{OpenapiComponent, OpenapiSchemaType}
 import sttp.tapir.codegen.openapi.models.OpenapiModels.OpenapiDocument
 import sttp.tapir.codegen.openapi.models.OpenapiSchemaType._
 import sttp.tapir.codegen.testutils.CompileCheckTestBase
 import sttp.tapir.codegen.util.DocUtils
+import sttp.tapir.codegen.validation.ValidationDefns
+import sttp.tapir.codegen.xml.XmlSerdeLib
 
 import scala.collection.mutable
 import scala.util.Try
@@ -471,7 +474,7 @@ class ClassDefinitionGeneratorSpec extends CompileCheckTestBase {
             validators = ValidationDefns.empty,
             generateValidators = true,
             packageReuse = PackageReuseContext.none,
-            seperateFilesForModels = false,
+            seperateFilesForModels = false
           )
           .endpointDecls(None)
     }
@@ -647,14 +650,16 @@ class ClassDefinitionGeneratorSpec extends CompileCheckTestBase {
   }
 
   it should "split models into separate files when seperateFilesForModels is true" in {
-    val result = new ClassDefinitionGenerator().classDefs(
-      TestHelpers.oneOfDocsWithMapping,
-      targetScala3 = isScala3,
-      jsonSerdeLib = JsonSerdeLib.Circe,
-      jsonParamRefs = Set("ReqWithVariants"),
-      fullModelPath = "foo.bar.models",
-      seperateFilesForModels = true
-    ).get
+    val result = new ClassDefinitionGenerator()
+      .classDefs(
+        TestHelpers.oneOfDocsWithMapping,
+        targetScala3 = isScala3,
+        jsonSerdeLib = JsonSerdeLib.Circe,
+        jsonParamRefs = Set("ReqWithVariants"),
+        fullModelPath = "foo.bar.models",
+        seperateFilesForModels = true
+      )
+      .get
 
     result.classReprs.keys should contain("")
     result.classReprs.keys should contain("ReqWithVariants")
